move pointer to separate file
